Sorts on user projects. not on day projects. Removed project create in project.repo
